+++ src/Rendering/SfmlWindowRenderTarget.hpp
@@ -25,12 +25,12 @@ namespace OsuReplayPlayer
 		bool isValid() const override;
 		sf::Vector2u getSize() const override;
 		void clear(sf::Color color) override;
-		void clear(sf::Vector2i pos, const sf::Image &image, sf::Vector2i newSize, unsigned char dimPercent) override;
+		void clear(sf::Vector2i pos, const sf::Image &image, sf::Vector2f newSize, unsigned char dimPercent) override;
 		void drawPixel(sf::Vector2i pos, sf::Color color) override;
 		void drawPoint(sf::Vector2f pos, sf::Color color) override;
 		void drawRectangle(sf::Vector2i pos, sf::Vector2u size, unsigned thickness, sf::Color color) override;
 		void drawFilledRectangle(sf::Vector2i pos, sf::Vector2u size, sf::Color color) override;
-		void drawImage(sf::Vector2i pos, const sf::Image &image, sf::Vector2i newSize, sf::Color tint, bool centered, float rotation) override;
+		void drawImage(sf::Vector2i pos, const sf::Image &image, sf::Vector2f newSize, sf::Color tint, bool centered, float rotation) override;
 		void drawCircle(unsigned thickness, sf::Vector2i pos, float radius, sf::Color color) override;
 		void drawFilledCircle(sf::Vector2i pos, float radius, sf::Color color) override;
 		void renderFrame() override;
